pyproject.toml
[project]
name = "sieve1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "cryptography>=46.0.5",
    "fastapi>=0.129.0",
    "gitpython>=3.1.46",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "tenacity>=9.1.4",
    "uvicorn>=0.40.0",
    "pyyaml>=6.0",
]
